flower_classification/requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astor==0.8.1
backports.weakref==1.0.post1
click==7.1.2
cryptography==2.1.4
enum34==1.1.10
exif-delete==0.9.5
Flask==1.1.2
funcsigs==1.0.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.30.0
gunicorn==19.10.0
h5py==2.10.0
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.16.6
piexif==1.1.3
Pillow==6.2.2
protobuf==3.12.2
pycrypto==2.6.1
pyxdg==0.25
PyYAML==5.3.1
scipy==1.2.3
SecretStorage==2.3.1
six==1.15.0
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow==2.0.0b1
tensorflow-estimator==1.14.0
tensorflow-hub==0.8.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
Werkzeug==1.0.1
wrapt==1.12.1
